packages/phi-detector/requirements.txt
transformers>=4.40.0
torch>=2.0.0
spacy>=3.7.0
pyyaml>=6.0
faker>=25.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
rich>=13.7.0
typer>=0.12.0
pydantic>=2.7.0
